backend/requirements.txt
fastapi
uvicorn
easyocr
python-multipart
opencv-python-headless
numpy
python-dotenv
supabase
